feat!: upgrade to tokio 1.0 (#664)
Upgrade Tokio to 1.0 as well as all related dependencies. In total, this
results in upgrading:

- async-tungstenite: 0.11
- hyper: 0.14
- hyper-rustls: 0.22
- hyper-tls: 0.5
- tokio: 1.0

The migration is not very interesting; changing
`tokio::time::delay_for`s to `tokio::time:sleep`s and such. Tokio has
changed the `rt-core` feature to `rt` and `rt-threaded` to
`rt-multi-thread`, so the examples have been updated for this.

Notably, due to `tokio::time::throttle` being moved to a separate crate
- `tokio-stream` - and working oddly, a thin throttle stream has been
manually implemented.

Closes #649.

Approved-by: Jens Reidel <adrian@travitia.xyz>
Approved-by: Valdemar Erk <valdemar@erk.io>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>