chore(http,lavalink): remove hyper's http1 feature (#670)
Remove the `http1` feature flag from hyper. It means that all
connections made will use `http2`, which is supported by Cloudflare and
Discord and has several advantages over `http1`, especially when the
same server is requested over and over again, which is the case here.

`http1` pulls in quite a bit of conditionally compiled code that won't
be used and might make the connections use `http1`, which isn't desired.
Both Discord/Cloudflare and Lavalink support `http2` and would benefit
from it, so forcing the use of it would only make sense. If a user would
like to keep using http1, they can add:

```toml
hyper = { version = 0.14, features = ["http1"] }
```

to their Cargo.toml to force enabling the feature.

Tested with the current http-proxy version and works fine with it. When
that is rewritten to 0.3, the `http1` server should be enabled.

Approved-by: BlackHoleFox <blackholefoxdev@gmail.com>
Approved-by: Vivian Hellyer <vivian@hellyer.dev>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Jens Reidel <adrian@travitia.xyz>